Move download to utils
